Added: permission check for products CRUD

+++ backend/src/products/products.controller.ts
@@ -1,14 +1,19 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, Query } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Query, Request, ForbiddenException} from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
+import { User, UserRole } from 'src/users/entities/user.entity';
 
 @Controller('products')
 export class ProductsController {
     constructor(private readonly productsService: ProductsService) { }
 
     @Post()
-    create(@Body() createProductDto: CreateProductDto) {
+    create(@Request() req, @Body() createProductDto: CreateProductDto) {
+      const user = req.user as User;
+      if (!user || req.user.role !== UserRole.ADMIN) {
+        throw new ForbiddenException('Access denied')
+      }
         return this.productsService.create(createProductDto);
     }
 
@@ -43,12 +48,20 @@ export class ProductsController {
     }
 
     @Patch(':id')
-    update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
+    update(@Request() req, @Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
+      const user = req.user as User;
+      if (!user || req.user.role !== UserRole.ADMIN) {
+        throw new ForbiddenException('Access denied')
+      }
         return this.productsService.update(id, updateProductDto);
     }
 
     @Delete(':id')
-    remove(@Param('id') id: string) {
+    remove(@Request() req, @Param('id') id: string) {  
+      const user = req.user as User;
+      if (!user || req.user.role !== UserRole.ADMIN) {
+        throw new ForbiddenException('Access denied')
+      }
         return this.productsService.remove(id);
     }
 }
